requirements.txt
kafka-python==2.0.2
Flask==2.0.3
requests==2.25.1
psycopg2-binary==2.9.3